+++ integration-tests/cli-test/dockerfiles/publish/publish.sh
@@ -15,11 +15,13 @@ npm publish -w @m2c2kit/assessment-cli-starter &
 npm publish -w @m2c2kit/assessment-color-dots &
 npm publish -w @m2c2kit/assessment-grid-memory &
 npm publish -w @m2c2kit/assessment-symbol-search &
+npm publish -w @m2c2kit/assessment-color-shapes &
 
 wait
 
 echo "npm publish to container registry complete: @m2c2kit/core, \
 @m2c2kit/addons, @m2c2kit/survey, @m2c2kit/sage-research, @m2c2kit/cli, \
 @m2c2kit/build-helpers, @m2c2kit/assessment-color-dots, \
+@m2c2kit/assessment-color-shapes, \
 @m2c2kit/assessment-grid-memory, @m2c2kit/assessment-symbol-search, \
 @m2c2kit/assessment-cli-starter"