+++ packages/assessments-demo/src/index.ts
@@ -7,6 +7,7 @@ import {
   EventType,
 } from "@m2c2kit/core";
 
+import { ColorShapes } from "@m2c2kit/assessment-color-shapes";
 import { ColorDots } from "@m2c2kit/assessment-color-dots";
 import { GridMemory } from "@m2c2kit/assessment-grid-memory";
 import { SymbolSearch } from "@m2c2kit/assessment-symbol-search";
@@ -14,14 +15,15 @@ import { CliStarter } from "@m2c2kit/assessment-cli-starter";
 import { Survey } from "@m2c2kit/survey";
 import { surveyJson } from "./surveyJson";
 
-const a1 = new ColorDots();
-const a2 = new GridMemory();
-const a3 = new SymbolSearch();
-const a4 = new CliStarter();
+const a1 = new ColorShapes();
+const a2 = new ColorDots();
+const a3 = new GridMemory();
+const a4 = new SymbolSearch();
+const a5 = new CliStarter();
 const survey = new Survey(surveyJson);
 
 const session = new Session({
-  activities: [a1, a2, a3, a4, survey],
+  activities: [a1, a2, a3, a4, a5, survey],
   canvasKitWasmUrl: "assets/canvaskit.wasm",
   sessionCallbacks: {
     /**